==> final_scripts/2_Model.Rmd <==
---
title: "Model and Analysis"
author: "Lawson et al"
date: "2023-06-19"
output: html_document
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
options(scipen=999)

library(tidyverse)
library("ggsci")
library(here)
library(knitr)
library(stargazer)

rm(list = ls())
```

# I. Load parameters
```{r, Load Params}
calibration = read.csv(here("data", "calibration_params.csv"))

for(i in 1:nrow(calibration)){
  assign(calibration$params[i], calibration$value[i])
}  


# Build the 2nd order parameters from parameter list
  # Differentiate parameters for Wild and Farmed inverse demands
alpha_w = alpha
alpha_f = alpha
beta_w = beta
beta_f = beta
  # Demand functions parameters
e   = gamma/(beta_w*beta_f - (gamma^2))
a_f = (alpha_f*beta_w - alpha_w*gamma)/(beta_w*beta_f - (gamma^2))
a_w = (alpha_w*beta_f - alpha_f*gamma)/(beta_w*beta_f - (gamma^2))
b_f = beta_f/(beta_w*beta_f - (gamma^2))
b_w = beta_w/(beta_w*beta_f - (gamma^2))

calibration %>% 
  select(-X)%>%
  mutate(value = round(value,2))%>%
  kable(row.names = F)
```

# II. Define functions
Define function where variable is either `x` i.e population stock, or `s` i.e price paid to poachers. All the parameters take default values specified in the global environment.  

* `growth(x, ...)` : logistic growth function, yields growth of population (in metric tons) 

* `pt_harvest(x,...)` : harvest (in metric tons) when trader is price taker

* `monop_harvest(x, ...)`: harvest (in metric tons) when trader is a monopolist

* `cournot_harvest(x, ...)`, `bertrand_harvest(x, ...)`: harvest (in metric tons) when trader and farmer compete in *Cournot*, i.e, set quantities strategically, and *Bertrand*, i.e, set prices strategically. 

* `price_poachers_cournot(x, ...)`, `price_poachers_bertrand(x, ...)`: price paid to poachers (in USD/metric tons) when trader and farmer compete in *Cournot* and *Bertrand*

* `cournot_farmed(s, ...)`, `bertrand_farmed(s, ...)`: quantity farmed (in metric tons) when trader and farmer compete in *Cournot* and *Bertrand*

*

```{r}
growth = function(x, 
                  k.=k, 
                  r.=r){
  y = r.*x*(1-x/k.)
  return(y)
}

# I. Price taker
pt_harvest = function(x, 
                      alpha. = alpha, 
                      beta. = beta, 
                      W. = W_mid, 
                      c. = c, 
                      sigma. = sigma){
  y = ((alpha.-c.) * x^2 * sigma.^2)/(4*W. + beta. * sigma.^2 * x^2)
  return(y)
}


# II. Monopoly
monop_harvest = function(x, 
                         alpha. = alpha, 
                         c. = c, 
                         sigma. = sigma, 
                         beta. = beta, 
                         W. = W_mid){
 y = ((alpha. - c.)*sigma.^2 * x^2) / (2*beta.* sigma.^2 * x^2 + 2*W.)
 return(y)
}

# III. Cournot

cournot_harvest = function(x, 
                             alpha_f. = alpha_f, 
                             alpha_w. = alpha_w, 
                             beta_f. = beta_f, 
                             beta_w. = beta_w, 
                             gamma. = gamma, 
                             sigma. = sigma, 
                             c. = c, 
                             v. = v, 
                             W. = W_mid){
  y = ((sigma.^2)*(x^2)*(2*beta_f.*(alpha_w. - c.) - gamma.*(alpha_f. - v.)))/(4*beta_f.*W. + (sigma.^2)*(x^2)*(4*beta_f.*beta_w. - (gamma.^2)))
  return(y)
}

price_poachers_cournot = function(x,
                                  alpha_f. = alpha_f, 
                                  alpha_w. = alpha_w, 
                                  beta_f. = beta_f, 
                                  beta_w. = beta_w, 
                                  gamma. = gamma, 
                                  sigma. = sigma, 
                                  c. = c, 
                                  v. = v, 
                                  W. = W_mid){
  y = (2*W.*(2*beta_f.*(alpha_w. - c.) - gamma.*(alpha_f. - v.)))/((sigma.^2)*(x^2)*(4*beta_f.*beta_w. - gamma.^2) + 4*beta_f.*W.)
  return(y)
}

cournot_farmed = function(s, 
                            alpha_f. = alpha_f, 
                            alpha_w. = alpha_w, 
                            beta_f. = beta_f, 
                            beta_w. = beta_w, 
                            gamma. = gamma, 
                            sigma. = sigma, 
                            c. = c, 
                            v. = v){
  y = (2 * beta_w. * (alpha_f. - v.) - gamma.* (alpha_w. - (s + c.)) )/(4*beta_w.*beta_f. - gamma.^2)
  return(y)
}

# IV. Bertrand

price_poachers_bertrand = function(x, 
                                   a_f. = a_f, 
                                   a_w. = a_w, 
                                   b_f. = b_f, 
                                   b_w. = b_w, 
                                   e.  = e, 
                                   c. = c, 
                                   v. = v, 
                                   sigma. = sigma, 
                                   W. = W_mid){
  y = (2*W.*b_w.*( b_f.*(2*a_w. + e.*v.) + c.*((e.^2) - 2*b_f.*b_w.) + e.*a_f.))/((sigma.^2)*(x^2)*(4*b_f.*b_w. - (e.^2)) + 2*W.*b_w.*(2*b_w.*b_f. - (e.^2)))
  return(y)
}

bertrand_harvest = function(x,
                              sigma. = sigma, 
                              a_f. = a_f, 
                              a_w. = a_w, 
                              b_f. = b_f, 
                              b_w. = b_w, 
                              e. = e, 
                              v. = v, 
                              c. = c, 
                              W. = W_mid){
  y = ((sigma.^2)*(x^2)*b_w.*(b_f.*(2*a_w.+e.*v.) + c.*((e.^2)- 2*b_f.*b_w.) + e.*a_f.))/((sigma.^2)*(x^2)*(4*b_w.*b_f. - (e.^2)) + 2*W.*b_w.*(2*b_f.*b_w. - (e.^2)))
  return(y)
}

bertrand_farmed = function(s, 
                             a_f. = a_f, 
                             a_w. = a_w, 
                             b_f. = b_f, 
                             b_w. = b_w, 
                             e. = e, 
                             c. = c, 
                             v. = v){
  y = b_f. *(2*b_w.*a_f. + v.*(e.^2 - 2*b_w.*b_f.) + e.*(a_w. + (s+c.)*b_w.))/(4*b_f.*b_w. - e.^2)
  return(y)
}

```

# III. Generate results

Generate results saved at `~/data/result1.csv`.
```{r, Results}

result1 = data.frame(x = seq(1,k))%>%
  mutate(growth = growth(x),
         # Price_taker
         pt_harvest = pt_harvest(x),
         ss_pt = growth - pt_harvest,
         # Monopoly
         monop_harvest = monop_harvest(x),
         ss_monop = growth - monop_harvest,
         # Cournot
         cournot_harvest = cournot_harvest(x),
         price_poachers_cournot = price_poachers_cournot(x),
         cournot_farmed = cournot_farmed(price_poachers_cournot),
         ss_cournot = growth - cournot_harvest,
         # Bertrand
         bertrand_harvest = bertrand_harvest(x),
         price_poachers_bertrand = price_poachers_bertrand(x),
         bertrand_farmed = bertrand_farmed(price_poachers_bertrand), 
         ss_bertrand = growth - bertrand_harvest
         )
head(result1)

# Save results
write_csv(result1, 
          here("data","result1.csv"))
```

Use `result1` to produce steady states:

```{r, Steady states}
steady_states = c()
  
for(variable in c('ss_monop', 'ss_cournot', 'ss_bertrand', 'ss_pt')){
  # Find where difference equations change sign
  ss = which(sign(eval(parse(text = paste0('result1$',variable)))) != lag(sign(eval(parse(text = paste0('result1$', variable))))))
  # Allow for up to 3 steady states
    if(length(ss)==1){
      steady_states = append(steady_states, c(result1[ss, 1], 
                                              NA,
                                              NA))
    }else if(length(ss)==2){
      steady_states = append(steady_states, c(result1[1, ss[1]], 
                                              result1[1, ss[2]],
                                              NA))
    }else if(length(ss)==3){
      steady_states = append(steady_states, c(result1[1, ss[1]], 
                                              result1[1, ss[2]],
                                              result1[2, ss[3]]))
    }
}



dat_monop = result1 %>% 
  subset(x %in% steady_states[1:3]) %>%
  select(x, monop_harvest)%>%
  mutate(q_farmed = NA,
         label = 'Monopoly')
# In case of cournot :
dat_cournot = result1 %>% 
  subset(x %in% steady_states[4:6])%>%
  select(x, cournot_harvest, cournot_farmed)%>%
  mutate(label = 'Cournot')
# In case of Bertrand : 
dat_bertrand = result1 %>% 
  subset(x %in% steady_states[7:9])%>%
  select(x, bertrand_harvest, bertrand_farmed)%>%
  mutate(label = 'Betrand')

dat_q_pt = result1 %>% 
  subset(x %in% steady_states[10:12])%>%
  select(x, pt_harvest)%>%
  mutate(q_farmed = NA, 
         label = 'Price taker')

names_ = c('Stock', 'Harvest', 'Farmed', 'Market structure')
colnames(dat_monop) = names_
colnames(dat_cournot) = names_
colnames(dat_q_pt) = names_
colnames(dat_bertrand) = names_

final_ = rbind(dat_q_pt, 
               dat_monop, 
               dat_cournot, 
               dat_bertrand)
final_ = final_[, c(4,1,2,3)]
kable(final_, row.names = F)
```

1. **Monopoly scenario** with cost parameter c=`r c` is a `r round((dat_monop$Stock - dat_q_pt$Stock)/dat_q_pt$Stock*100,2)`, % variation in *population stock* in tonnes (compared to OA) or, equivalently, of `r round((dat_monop$Stock - dat_q_pt$Stock)*1000/25.7)` adult totoabas

2. **Monopoly scenario** with cost parameter c= `r c` is a difference of `r round(dat_q_pt$Harvest-dat_monop$Harvest ,2)`(in tons) in *harvest* or, equivalently, of `r round((dat_q_pt$Harvest-dat_monop$Harvest)*1000/25.7)` adult totoabas

3. **Worst case scenario** with cost parameter c=`r c` is a `r round((dat_bertrand$Stock - dat_monop$Stock)/dat_monop$Stock*100,2)` % variation in *population stock* in tonnes or, equivalently, of `r round((dat_bertrand$Stock - dat_monop$Stock)*1000/25.7)` adult totoabas

4. **Worst case scenario** with cost parameter c=`r c` is a difference of `r round(dat_bertrand$Harvest- dat_monop$Harvest,2)` (tons) in *harvest* or, equivalently, of `r round((dat_monop$Harvest-dat_bertrand$Harvest)*1000/25.7,2)` adult totoabas


5. **Best case scenario** with cost parameter c=`r c` is a `r round((dat_cournot$Stock - dat_monop$Stock)/dat_monop$Stock*100,2)` % variation in *population stock* in tonnes or, equivalently, of `r round((dat_cournot$Stock - dat_monop$Stock)*1000/25.7,2)` adult totoabas

6. **Best case scenario** with cost parameter c=`r c` is a difference of `r round(dat_cournot$Harvest- dat_monop$Harvest,2)` (in tons) in harvest in tonnes or, equivalently, of `r  round((dat_monop$Harvest-dat_cournot$Harvest)*1000/25.7,2)` adult totoabas

```{r, Graphic output}
actual_steady = steady_states[!is.na(steady_states)]

lines_ = F

plot_ = result1 %>%
  subset(x < 1.5*max(actual_steady))%>%
  ggplot(aes(x=x))+
  geom_line(aes(y = growth, colour = 'Growth'), linewidth = 1)+
  geom_line(aes(y = monop_harvest, colour = 'Monopoly'), linewidth = 1)+
  geom_line(aes(y = cournot_harvest, colour = 'Cournot'), linewidth = 1)+
  geom_line(aes(y = bertrand_harvest, colour = 'Bertrand'), linewidth = 1)+
  geom_line(aes(y = pt_harvest, colour = 'Price taker'), linewidth = 1) +
  theme_bw()+
  ylab('Growth / harvest')+
  xlab('Population stock')+
  scale_color_aaas()+
  labs(colour = ' ')+
  theme(legend.position = 'bottom')

if(lines_ == T){
  plot_ = plot_ + 
    # Vertical lines
  geom_segment(x = actual_steady[1], xend = actual_steady[1], y=0, yend=growth(actual_steady[1]), linetype = 2)+
  geom_segment(x = actual_steady[2], xend = actual_steady[2], y=0, yend=growth(actual_steady[2]), linetype = 2)+
  geom_segment(x = actual_steady[3], xend = actual_steady[3], y=0, yend=growth(actual_steady[3]), linetype = 2)+
  geom_segment(x = actual_steady[4], xend = actual_steady[4], y=0, yend=growth(actual_steady[4]), linetype = 2)+

  # Horizontal lines
  geom_segment(x = 0, xend = actual_steady[1], y=growth(actual_steady[1]), yend=growth(actual_steady[1]), linetype = 2)+
  geom_segment(x = 0, xend = actual_steady[2], y=growth(actual_steady[2]), yend=growth(actual_steady[2]), linetype = 2)+
  geom_segment(x = 0, xend = actual_steady[3], y=growth(actual_steady[3]), yend=growth(actual_steady[3]), linetype = 2)+
  geom_segment(x = 0, xend = actual_steady[4], y=growth(actual_steady[4]), yend=growth(actual_steady[4]), linetype = 2)
}
plot_ 
ggsave(here('visuals', 'Figure3.jpg'))
```
